move fixtures out of test file
